update makefile for version.h
git-svn-id: svn+ssh://140.110.240.196/partclone_svn@305 93b75dfe-da2d-4241-a08d-2141769abc04

+++ partclone/src/fatclone.c
@@ -178,8 +178,8 @@ extern int check_fat_status(){
 /// mark reserved sectors as used
 static unsigned long long mark_reserved_sectors(char* fat_bitmap, unsigned long long block)
 {
-    int i = 0;
-    int j = 0;
+    unsigned long long i = 0;
+    unsigned long long j = 0;
     unsigned long long sec_per_fat = 0;
     unsigned long long root_sec = 0;
     sec_per_fat = get_sec_per_fat();
@@ -235,7 +235,7 @@ unsigned long long check_fat32_entry(char* fat_bitmap, unsigned long long block,
 {
     uint32_t Fat32_Entry = 0;
     int rd = 0;
-    int i = 0, j = 0;
+    unsigned long long i = 0, j = 0;
 
     rd = read(ret, &Fat32_Entry, sizeof(Fat32_Entry));
     if (rd == -1)
@@ -263,7 +263,7 @@ unsigned long long check_fat16_entry(char* fat_bitmap, unsigned long long block,
 {
     uint16_t Fat16_Entry = 0;
     int rd = 0;
-    int i = 0, j = 0;
+    unsigned long long i = 0, j = 0;
     rd = read(ret, &Fat16_Entry, sizeof(Fat16_Entry));
     if (rd == -1)
 	log_mesg(2, 0, 0, 2, "read Fat16_Entry error\n");
@@ -290,7 +290,7 @@ unsigned long long check_fat12_entry(char* fat_bitmap, unsigned long long block,
     uint16_t Fat16_Entry = 0;
     uint16_t Fat12_Entry = 0;
     int rd = 0;
-    int i = 0, j = 0;
+    unsigned long long i = 0, j = 0;
     rd = read(ret, &Fat16_Entry, sizeof(Fat16_Entry));
     if (rd == -1)
 	log_mesg(2, 0, 0, 2, "read Fat12_Entry error\n");
@@ -350,12 +350,12 @@ extern void initial_image_hdr(char* device, image_head* image_hdr)
 /// readbitmap - read and check bitmap
 extern void readbitmap(char* device, image_head image_hdr, char* bitmap, int pui)
 {
-    int i = 0, j = 0;
+    unsigned long long i = 0, j = 0;
     int rd = 0;
     unsigned long long block = 0, bfree = 0, bused = 0, DamagedClusters = 0;
     unsigned long long cluster_count = 0;
     unsigned long long total_sector = 0;
-    int FatReservedBytes = 0;
+    unsigned long long FatReservedBytes = 0;
     uint16_t Fat16_Entry = 0;
     uint32_t Fat32_Entry = 0;
     extern cmd_opt opt;
@@ -414,7 +414,7 @@ extern void readbitmap(char* device, image_head image_hdr, char* bitmap, int pui
 /// get_used_block - get FAT used blocks
 static unsigned long long get_used_block()
 {
-    int i = 0;
+    unsigned long long i = 0;
     int rd = 0;
     unsigned long long block = 0, bfree = 0, bused = 0, DamagedClusters = 0;
     unsigned long long cluster_count = 0, total_sector = 0;
